less-than: implemented remaining less than operators
TESTS: verify that less than of an integer and a double, a double and
an integer and two integers is executed correctly.

The execute functions for the remaining less than operators were
implemented and their function pointers replacing the null pointer in
the corresponding code instance one at a time.