greater-than: implemented the greater than operators
TESTS: verify that greater than of two doubles, an integer and a double,
a double and an integer and two integers is executed correctly.

The execute functions for the greater than operators were implemented
and their function pointers replacing the null pointer in the
corresponding code instance one at a time.