less-equal: implemented compiling of the less than or equal operator
TEST: verify that a less than or equal expression is compiled to the
end.

The less than or equal operator contains two character (a less than and
an equal character).  The compile class get comparison operator codes
function needed to be modified to handle the possibility of a two
character operator.

To accomplish this, the get comparison operator codes will first add the
next peek character to a blank keyword string variable, and see if there
are codes for the character.  If there are, it could be a one character
operator or a two character operator.  After consuming the first
character, the next peek character is added to the keyword string.  If
there are codes for the two character keyword, then the second character
is consumed and the second set of codes is returned, otherwise the first
set of codes is return.  A null pointer is returned if there were no
codes found.

The function with this additional logic was getting larger than desired
and so some refactoring was done.  The codes searching code was spun off
into a new check for comparison operator codes helper function operator
function class, that takes a reference to the compile and contains a
keyword string member variable.  Besides the function operator function,
it contains a private codes with next peek character function that adds
the next peek character to the keyword and finds codes using the
precedence comparison operator data function for the keyword.  If found,
then the next character is consumed.  This was put into a function
because it needs to be repeated twice for up to two characters.

The four less than or equal code instances were added (with all null
pointers) along with the less than or equal numeric operator codes
instance with their precedence level and two character keyword.

Some other minor refactoring was done with the other compiler functions
so that they were all consistent with one another.  A new get alpha
only word helper function was added for the get keyword function.