less-equal: implemented execution of the less than or equal to operators
TESTS: verify that less than or equal to of two doubles, an integer and
a double, a double and an integer and two integers is executed
correctly.

The execute functions for the less than or equal to operators were
implemented and their function pointers replacing the null pointer in
the corresponding code instance one at a time.