relational: refactored relational operator execute functions
The execute functions for each additional relational operator that was
implemented was a copy, paste and modify affair.  This is duplicated
code and it needed to be eliminated.

This was accomplished by creating a set of four template execute
functions taking a compare function as a template argument.  For each
operator, two inline compare functions were implemented taking either
two double or two integer arguments, which compare the arguments with
the appropriate operator and returns the boolean result.  The execute
function pointer for each code instance was then replaced with the
template function and the appropriate compare function (the correct
overload compare function, either double or integer arguments, is
selected based on the arguments passed by the execute function).