equal: adding parsing of the equal to operator
TEST: verify that a simple equal to expression is parsed to the end of
the expression.

The equal to operator is at a new equality precedence that will be
shared with the not equal to operator.

The four equal to codes instances were added (with all null function
pointers) along with the equal to numeric operator codes instance with
their precedence level and keyword (an equal character).

These instance were no sufficient to get the test to pass.  The reason
was that the equal operator was not considered a comparison operator
because it was added to the comparison operator data in the precedence
info class.  To get this to occur, the precedence info add operator data
function was modified to add both relation and equality precedence
operators to the comparison operator data map.

The test now passed, but the equal operator was being compiled by the
compile relation function.  Since this operator is at a different
precedence level it needs to be compiled by a separate compile equality
function.  But that is it for this commit since this tests passed.