equal: compile equal to operator at correct precedence level
TEST: verify that the equal to operator is compiled at the correct
precedence level below relational operators.

This test contains the "0=1<2" expression, which should have the less
than operator compiled first followed by the equal to operator.  This
test initially failed because equal to was compiled at the same
precedence as less than, so it was incorrectly compiled first.

Because the not equal operator (not yet implemented) needs to be parsed
at the same time as the less than and less than or equal operator (since
it shares the same first character), the equality operators will be
parsed when the compile relation function is executed.  However, if the
operator is an equality operator, it can't be compiled until the
compile equality function is executed.  The compiler therefore needs to
cached the equality codes, return null codes to the compile relation
function, but return these equality codes to the compile equality
function.

An equality codes pointer was added to the compiler class along with a
private access function to set them.  The check for comparison operator
codes helper class was made a friend class of the compiler class so that
it can access this private function.

The compiler get comparison operator codes function was given a
precedence argument so that it can take the appropriate action depending
on the type of comparison operators found, which is returning relation
operator codes is the relation precedence is specified or caching
equality operator codes, or returning the cached equality codes if any
when the equality precedence is specified (which is done at the
beginning of the function).

The check for comparison operator codes codes with next peek character
function was modified return the comparison operator codes if they have
a relation precedence (this function will not be called for an equality
precedence), or cache them if they are equality precedence.

Since the comparison operator data map in the precedence info class now
needs to store the precedence with the comparison operator codes
pointer, a new comparison data class was added to hold the codes pointer
and the precedence level.  An instance to one of these is now returned
by the comparison operator data functions.

The compile equality function was implemented in the expression compiler
class that is similar the compile relation function - gets the
comparison operator codes for an equality precedence, and it not null,
compilers the second relation expression operand, selects the proper
code from the operand data types, and adds the instruction to the code.

The code was refactored a bit which the check for comparison operator
codes class was shortened to comparison operator codes (this name with
the member function names were getting to long).  The code for returning
either the relation operator codes or caching the equality codes was
extracted into a new relation codes function in this class.