equal: return integer data type for equal to operator
TEST: verify that the equal to returns the integer data type.

This test (double-double equal to operator) initially failed because
the compile relation function was just returning the left hand side
data type, which was double for this test.

To make this test pass, instead of setting the left hand side to the
operator info return data type, which would be incorrect if any operand
data type is double, the left hand side data type was just set to the
integer data type.