equal: allow white space after the operator
TEST: verify that a equal to operator with white space after is
compiled to the end of the expression.

It was initially puzzling that this test was failing since the equal to
operator was being processed by the same function that the relational
operators which was handling white space after the operator.

The issue was that the get comparison operator codes was now checking
equality precedence and then returning any saved equality operator
codes.  This check was before skipping white space after the operator,
so white space was only being skipped after relational operators.

To make this test pass, the call to skip white space was moved to after
a comparison operator is found, but before determining if the operator
codes found was a relational operator so now white space is skipped
after any comparison operator.

The get comparison operator codes function was refactored by extracting
the code returning saved equality codes into a new saved equality
operators codes function.  The get comparison operator codes function is
now nicely parallel between handling of equality and relational operator
codes.