equal: implemented execution of the equal to operators
TESTS: verify that equal to of two doubles, an integer and a double, a
double and an integer and two integers is executed correctly.

Making use of the previously implemented execute compare template
functions, equal functions were added to compare equality between either
two doubles or two integers, and a pointer these functions were used as
the template argument of the execute compare template functions.

OTHER: all of the little compare functions were changed to inline
functions.