operators: added missing checks for missing left side operand
TESTS: verify that when there is no left hand side operand (expression
starts with the operator) that an "expected numeric error" occurs.

This flaw was discovered when adding the not equal operator.  These
changes were stashed so that the flaw could be corrected.  The flaw
occurred because none of the compile expression functions (except for
exponential and negation) were checking if a null data was returned
after compiling the left hand side operand.  It was erroneously assumed
that only the compile exponential function needed to check this.

Each test and correction was added one at a time to the compile product,
integer division, modulo, summation and relation functions.  A check was
not needed in the compile equality function because in the compile
relation function, if the data type is null from the left operand, the
equality codes is not saved and are not there when requested in the
compile equality function, and the unmodified null left hand side
operand data type is returned.