operators: added missing checks for bad right side operand
TESTS: verify that when there is a bad right hand side operand
(say the operand starts with another operator) that an "expected
numeric error" occurs.

This flaw was also discovered when adding the not equal operator.  The
flaw occurred because none of the compile expression functions (except
for exponential) were checking if a null data was returned after
compiling the right hand side operand.  It was erroneously assumed that
only the compile exponential function needed to check this.

Each test and correction was added one at a time to the compile product,
integer division, modulo, summation, relation and equality functions.