not-equal: adding parsing of the not equal to operator
TEST: verify that a simple not equal to expression is parsed to the end
of the expression.

The not equal to operator is at a same equality precedence that will be
shared with the equal to operator.

The four not equal to codes instances were added (with all null function
pointers) along with the not equal to numeric operator codes instance
with their precedence level and keyword (an equal character).

Adding these instances were not sufficient to get the test to pass.  The
test still failed because the first (less than) character of the not
equal to operator was grabbed as a relation operator, but when compiling
its operand, returned an "expected numeric expression" at the second
(greater than) character of the not equal to operator.

To get this not to occur, the local one character codes variable in the
comparison operator codes help class function operator function was made
a member variable of this class.  Then in the relation codes function
when the equality codes are set (upon finding a two character operator
codes), the one character codes pointer is set to null to prevent the
function operator function from returning the one character less than
codes.  The test now passed.