not-equal: implemented execution of the not equal to operators
TESTS: verify that not equal to of two doubles, an integer and a double,
a double and an integer and two integers is executed correctly.

Making use of the previously implemented execute compare template
functions, not equal functions were added to compare inequality between
either two doubles or two integers, and a pointer these functions were
used as the template argument of the execute compare template functions.