not: adding parsing of the not operator
TEST: verify that a simple not expression is parsed to the end of the
expression.

The not operator is at its own not precedence level.  The not operator
only has a single code taking an integer operand.  A new not operators
codes class was added, which contains only a single code.  The select
function simply returns the single code.

The single not code instance was added (with null function pointers)
along with the not operator codes instance with its precedence level and
keyword.

Since the not operator is at its own precedence level (and that it is a
unary operator), a new compile not function was added to the expression
compiler class.  The compile numeric expression was changed to call the
compile not function instead of the compile equality function.  To
start, the compile not function first calls the get word operator codes
for the not precedence, then calls the compile equality and returns the
data type returned.  This was sufficient to get the new test to pass.

However, one of the expression tests were now failing - the test that
verifies an error is thrown for an empty input stream.  The error was
still thrown, but the column was incorrect.  The column should have been
zero, but was the equivalent of -1 indicating that the input stream had
incorrectly been read past the end of the stream.

The problem was that the skip white space function (which uses the
standard skip white space extractor function), was now being called at
the beginning of the expression by the get keyword function via the
get word operator codes function.  If white space is extracted from a
stream with no characters remaining, it puts the stream into an error
condition.

The first attempt to fix this problem was to replace the use of the
white space stream extractor with a simple while loop that calls the
peek next character function (which gets the column before peeking the
next character that will put the stream into an error condition if there
is no character to peek), and if the character is white space, to get
the white space character and peek for the next character.

The empty input stream test passed, but now other tests were failing -
tests with print statements.  The problem was that the peek next
character function returns the first character of the word if the word
is not empty, and the get keyword function calls the skip white space
function after getting and setting the word and therefore the peek next
character function was called.  So the first character  of the word was
being returned and white space was now not being skipped after the
command (print) keyword.

This problem was that the word member can't be set until after the white
space after the keyword is skipped.  To correct this, the skip white
space function call was moved from the get keyword function to the get
alpha only word function that gets the keyword into a local variable
that is returned.