not: return integer data type for not operator
TEST: verify that the not returns the integer data type.

This test (a double constant after the not operator) initially failed
because the compile not function was just returning the data type of the
operand, which was double for this test.

To make this test pass, instead just returning the data type of the
operand, the data type is set to integer if there was a not operator.