not: compile multiple not operators
TEST: verify that multiple not operators are compiled to the end of the
expression.

This test initially failed because the compile not function was only
looking for one not operator.

The test was made to pass by restructuring the compile not function.
The assignment of the codes variable was put into the if statement.  If
the codes are set, the compile not function is called recursively.  The
code is then selected, added to the program code, and the integer data
type is returned.  If the codes are not set, then the compile equality
function is called and its return data type is returned directly.