not: recreate expressions with not operator
TEST: verify the not operator is recreated correctly.

The pointer to the recreate unary operator function was added to the not
code instance.  However, since the not operator consists of letters,
there needs to be a space after the not operator.  The append unary
operator helper function of the recreator class was modified to append
a space if the last character of the operator is an alphabetic character
(and this won't interfere with the negate operator).

Note that a space is not needed before the not operator since a not
operator will either be at the beginning of an expression (the
preceding command will have a space after it) or will be after either
another not operator (strange, but possible) or a binary operator, both
of which will have spaces after them.