not: change double constant operand to an integer constant
TEST: verify when there is a double operand after the not operator, that
the double constant is changed to an integer constant.

Since constants contain both their integer and double values in the
dictionary, either can be used as needed.  This saves time by avoiding
an unnecessary conversion during program execution.  Even though a
constant was interpreted as an double constant (a decimal point or an
exponent), its value was converted to an integer and stored in the
dictionary.  All that needs to be done is that the constant double code
put into the program needs to be changed to a constant integer code.

This test initially failed because the execute not function assumed that
its operand had pushed an integer value on top of the value stack, but
for an double constant, the double value was set, and the value was not
interpreted correctly giving the wrong result.

The compile not function in the expression compiler class was changed to
call a new convert to integer function in the compiler class.  To make
this test pass, all this new function needed to do was to get the offset
where the constant code was added to the code.  The constant double was
just added, so its code and operand are the last two words in the code
line, therefore its offset is two less than the size of the code line.
The word at this offset is set to the constant integer code.  If the
code is already a constant integer, this still works.