not: only change to integer constant if operand was a constant
TEST: verify that a non-constant code is not erroneously changed to an
integer constant with a not operator.

If the code of the previous operand is not a constant, code will be
corrupted.  This is worsened if the last operand was an operator as the
convert to integer could be changing an operand to constant integer
code.

To get this test to pass, the last constant is only changed to an
integer constant if the last constant was constant flag is true.