not: add convert to integer code if operand is double
TEST: verify that the internal convert to integer code is added to the
code if for a non-constant double operand.

The special convert to integer will be inserted into the program for the
not operator and later for functions meant to take a only an integer so
that many codes with don't need to be implemented.  The exception being
the math operators that that could be used with either type (otherwise
these convert to double and integer codes would be in code in many
places, which could effect execution time, and requiring the user to use
the CVTDBL and CVTINT functions to avoid this is an undesirable burden).

To get this test to pass required a number of changes.  A convert to
integer code instance (with null pointers for the recreate and execute
functions) was added.  The convert to integer function was given an
operand data type argument, which is now checked to see if a conversion
to integer is needed (the operand data type was added to the caller).
If the last operand was not a constant, then the internal convert to
integer code is added to the program.