not: produce error if double constant cannot be changed to integer
TEST: verify that a compile error is produced if the operand of a not
operator is a double constant that cannot be converted to an integer
constant because it is out of the range of an integer.

Several changes were needed to get this test to pass, which were done in
a single commit.  A check was added to the convert to integer function
to see if the double constant was within the range of an integer before
changing to an integer constant, otherwise an error is thrown.

This check starts with obtaining the operand of the constant within the
program code (which is the word after the instruction code of the
constant).  This operand is passed to a new constant number dictionary
convertible to integer function, which checks if the double value of the
constant is within the range of an integer.  The check integer overflow
function already contained this check, but the throwing of the error was
not needed, so this check was extracted into a new within integer range
template function.

To produce the error pointing to the constant, the column and length of
the constant was needed to the error.  Last constant column and length
members were added to the compiler class.  These are set by the add
numeric constant instruction function, which required a new column
argument (the length can be obtained from the number string argument).
A value of this argument was added to the constant number compiler
class, which has this value its first column member.  The other callers
were various unit tests, and since these tests don't make use of an
error column, a value of zero was added for the value of this argument.

Upon seeing the final convert to integer function was getting rather
large, it was refactoring into two smaller functions to better clarify
its intent.  The convert to double function was similarly refactored
with one additional smaller function.