not: corrected problem with extra spaces recreated after operator
TEST: verify that an expression with a positive constant after a not
operator is recreated correctly.

This test initially failed because a space always output after a unary
operator that ends with a letter, but another space was also output
before a constant after a unary operator.  This space before a constant
was necessary so that a negate before a positive constant would be
correctly parsed as a negate operator and constant and not a negative
constant.

This problem was corrected by only appended a space before a constant if
the recreated string so far (top string on the recreator stack) does not
already have a space.