not: recreate convert to integer code to nothing
TEST: verify that an internal convert to integer code is recreated to
nothing.

To make this test pass, and a pointer to the recreate nothing function
replaces the null pointer in the convert to instance code instance.

OTHER: the convert to double and integer code instances were renamed
with an 'internal' prefix, necessary so to not conflict with the yet to
be implemented CVTDBL and CVTINT functions.