not: execute the internal convert to integer code
TEST: verify that an expression requiring an internal convert to integer
code is executed correctly.

An execute convert integer function was implemented that converts the
double on top of the stack to a integer and puts it into the integer
value on top of the stack.  A pointer to this function replaces the null
pointer in the convert to integer code instance.

Since the functionality of converting a double to an integer and putting
it in the top item of the executer stack, a new set top integer from
double function was added to the executer class and the two callers
(execute convert to integer and execute integer divide) changed to call
this new function.