not: check for overflow when executing internal convert to integer
TEST: verify that an overflow run error is produced if the double being
converted to integer is not within the range of an integer.

Note that the entire operand should be indicated as having the error an
not the not operator.  This will be more appropriate for functions (for
example the second or third arguments of the MID$ function) when an
overflow occurs with an argument so that the proper argument is
indicated to where the overflow occurred.

To make this test pass, a call to the check integer overflow function
was added to the execute convert to integer function with the operand
(the double that is on top of the stack).

This was not sufficient to make the test pass, specifically the code
crashed because there were no error indicated put into the recreated
string (since the recreate function of the convert to integer code, the
offset in the run error where the error occurred, did nothing).

To get this test to pass, a mark operand if error function was added to
the recreator class, which gets the top string (the operand) and then
appends a start error indicator if at the error offset, appends the
operand string, and appends an end error indicator if at the error
offset.  If not at the error offset, the net effect of this function is
nothing.

A recreate convert to integer function was added that calls the new
recreator mark operand if error function, and its function pointer
replaced the recreate nothing function pointer.

OTHER: the convert to double and integer code instances were renamed
back removing the 'internal' prefix, not necessary because the BASIC
functions are named CDBL and CINT so there is no name conflict.