and: added parsing of the and operator
TEST: verify that a simple and expression is parsed to the end of the
expression.

The AND operator is at its own and precedence level.  The AND operator
has only a single code taking integer operands.  A new logic operator
codes class was added, which contains only a single code (taking two
integer operands).  The select function simply returns the single code.

The single and code instance was added (initially with null function
pointers) along with the and operator codes instance with its precedence
and keyword.

Since the AND operator is at its own precedence level, a new compile and
function was added to the expression compiler class.  The compile
numeric expression was changed to call the compile and function instead
of the compile not function.  To start, the compile and function first
calls the compile not function for the left side operand and stores the
returning data type.  It then calls the get word operator codes for the
and precedence.  If codes are returned, the compile not function is
called again for the right side operand.  The data type is returned.

These changes were not sufficient to get the test to pass, in fact
several test failed because they got stuck in an infinite loop.  The
problem was caused in the get keyword function.  The AND operator is
not the first to call this function for a binary operator, as the MOD
operator first called it, and the word was saved since it was not the
a MOD keyword.  When the compile and called it again, since the AND
keyword was already read, the get keyword function continued trying to
read a word at the right side operand.

To correct this, a check was added to the get keyword function to only
skip white space and read a word if the saved word variable is empty.

Note that the logic operator codes select function is not covered by
any of the unit tests (yet).  Normally uncovered code should not be
added until covered by a test, but since the select function is virtual,
the code will not compile without it.