and: return integer data type for and operator
TEST: verify that the and returns the integer data type.

This test (a double constant before the and operator) initially failed
because the compile and function was just returning the data type of the
left side operand, which was double in this test.

To make this test pass, instead just returning the data type of the
left side operand, the data type is set to integer if there was an AND
operator.