and: make sure left side operand is an integer
TEST: verify that a double left side constant is converted to an integer
constant.

This test initially failed when the value of a double left side operand
was interpreted as an integer.  To make the test pass, in the compile
and function after getting AND operator codes, a call is made to the
compiler convert to integer function with the left side operand data
type.

No further tests are needed for the left side operand since the convert
to integer function has already been fully tested (changed integer
constants, add an internal convert to integer code, and check for a
constant double out or range error and run time out of range error).