and: make sure right side operand is an integer
TEST: verify that a double right side constant is converted to an
integer constant.

This test initially failed when the value of a double right side operand
was interpreted as an integer.  To make the test pass, in the compile
and function, the return data type from the call to the compile not
function for the right side operand is passed to a second convert to
integer function call.