expression: implemented generic compile logic operator function
The expression compiler compile functions for the rest of the logic
operators will follow the same form as the compile and function.  So
instead of continuing to duplicate this pattern (the other math
operators are similar but without the convert to integer calls),
compile logic operator function was added taking a precedence level
and a member function pointer to a compile function, the only info that
differs between each of the logic operators.

The code in the compile and function was replaced with a call to this
new function passed the and precedence and a member function pointer to
the compile not function.