expression: implemented generic compile more operator functions
Following the pattern of the compile logic operator function, similar
compile math and comparison functions were implemented and code in the
compile equality, relation, summation, product, and exponential
functions was replaced with a call to one of these functions.

The compile modulo and integer division couldn't be modified as they
deviate from the pattern (modulo is a word operand, integer division
does calls to convert to double).  The new compile operator functions
are also very similar to each other, so more refactoring is needed to
reduce the duplicate code.