expression: implemented a single generic compile operator function
After the previous changes, even though some of the duplication was
removed, there were still three generic functions that were very similar
to each other.  Therefore, through a series of small refactorings, these
were reduced to a single function.

One of the differences was between numeric operators and comparison
operators where comparison operators return an integer data type only.
To eliminate this difference, a new comparison operator codes class was
added, which uses the numeric operator codes as a base with only the
select function reimplemented.  The select of this new class calls the
select function of the numeric operator codes base and changes the data
type to integer before returning.  The comparison codes instances were
then changed to this new class.

The final compile operator function takes four arguments, a member
function pointer to the function to call for compiling operands, the
precedence level of the operator to compile, a function pointer of the
function to call for getting the operator codes, and a function pointer
of the function to call for converting the operands to the needed type
if necessary.

Three get codes functions were implemented, which take a reference to
the compiler instance and the precedence level.  These functions call
the get symbol, comparison or word operator codes function.

Three convert functions were implemented, which take a reference to the
compiler instance and the data type of the operand.  The no convert
function does nothing and is the default.  The convert to double and
integer functions call the same function of the compiler instance.

To simplify the compiler operator function a bit, some of the code was
extracted to a new add selected code function that handles selecting
the code from the data type of the operands, add the instruction code of
the selected code and returns the result data type of the code.