or: added parsing of the or operator
TEST: verify that a simple or expression is parsed to the end of the
expression.

The OR operator is at its own or precedence level.  The OR operator
has only a single code taking integer operands.

The single or code instance was added (initially with null function
pointers) along with the or operator codes instance with its precedence
and keyword.

Since the OR operator is at its own precedence level, a new compile or
function was added to the expression compiler class.  The compile
numeric expression was changed to call the compile or function instead
of the compile and function.  Using the just implemented compile
operator function with arguments of the or precedence, the compile and
function for the compile operand function, word get codes function, and
convert to integer function.

To make the calls to the compile operator function clearer, the
precedence argument was made first swapping with the compile operand
member function pointer.  Putting the precedence first better clarifies
the operator being compiled, and what operator(s) are the next higher
operator(s) to compile.