xor: added parsing of the xor operator
TEST: verify that a simple XOR expression is parsed to the end of the
expression.

The XOR operator is at its own or precedence level.
The XOR operator has only a single code taking integer operands.

The single xor code instance was added (initially with null function
pointers) along with the xor operator codes instance with its precedence
and keyword.

Since the XOR operator is at its own precedence level between AND and
OR, a new compile or function was added to the expression compiler
class.  The compile or function was changed to have the compile xor
function as the compile operand function instead of the compile and
function.  Calling the compile operator function with arguments of the
xor precedence, the compile and function for the compile operand
function, word get codes function, and convert to integer function.

No further compile tests are needed since the compile operator function
has already been fully tested.