eqv: implemented execution of the equivalence operator
TESTS: verify that EQV of integers is executed correctly.

The execute eqv function was implemented and its function pointer
replaced the null pointer in the and code instance.  The equivalence
operator is the same as exlusive-nor, the this function uses the C++ '^'
bit-wise xor operator of the operands and C++ '~' 1's compliment unary
operator on the result.