imp: implemented execution of the implication operator
TESTS: verify that IMP of integers is executed correctly.

The execute imp function was implemented and its function pointer
replaced the null pointer in the imp code instance.  The implication
operator is the same as not of the left hand side value ored with the
right hand side value.  This function uses the C++ '~' 1's compliment
unary operator and the c++ '|' or operator.

Upon testing, an unexpected result occurred.  The problem was caused
due to the operands being reversed.  All of the logic operators
incorrectly labeled left hand side.  This did not affect the results of
the other operators because they are commutative, but the IMP operator
is not, which is why the result did not match what was expected.

This was corrected in all of the logic operator execute functions by
renaming the local variable the right hand side value, and reversing
the operands in the calculations (for correctness).