expression: corrected problem with recreation of unary operators
TEST: verify that the higher precedence negate operator on the left side
operand of lower precedence add operator is recreated correctly.

This test was initially failing because parentheses were incorrectly
added around the left side operand because of a check that was always
adding parentheses on the left side operand when it contained a unary
operand.  This was necessary for the higher precedence exponential
operator, but not correct for lower precedence operators.

To correct this problem, the precedence of the unary operator on the
left side operand needs to be checked against the binary operator
precedence and parentheses added only if the unary operator precedence
is lower than the binary operator.

To accomplish this, the unary operator flag on the recreator stack item
was changed to a unary operator precedence.  The stack item member and
access functions were changed accordingly.  In the recreate binary
operator, the right side unary operator precedence is simply transferred
to the top stack item unary operator precedence.

The default value of the unary operator precedence is set to the
precedence level of operand (the highest).  This will prevent
parentheses from being added for non-unary left side operands.  So that
the check for a right side unary operator remains clear in the append
right operand function, an is unary operator function was added to the
stack item class (which checks if the unary operator precedence is not
the default operand precedence level).

OTHER: Changed the precedence level enumeration to an enumeration class,
which was possible since the enumerators were no longer being used as
indexes.  Renamed the top string function of the recreator class to the
move top string function to better clarify that it moves the top string
from the top item of the stack.  Made several recreator function private
since they were only being called internally by the recreator functions.
The two recreator stack item member functions were moved to the
recreator source file since the stack item class is private to the
recreator class.